task 10 laravel

+++ database/migrations/2023_11_17_175728_create_cars_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('cars', function (Blueprint $table) {
             $table->id();
             $table->string('carTitle',100);
-            $table->float('price');
+            $table->double('price');
             $table->text('description');
             $table->boolean('published');
             $table->string('image',100);
